Titan followup changes
