+++ db/dbformat.h
@@ -53,8 +53,15 @@ enum ValueType : unsigned char {
   kTypeNoop = 0xD,                        // WAL only.
   kTypeColumnFamilyRangeDeletion = 0xE,   // WAL only.
   kTypeRangeDeletion = 0xF,               // meta block
-  kTypeColumnFamilyBlobIndex = 0x10,      // Blob DB only
-  kTypeTitanBlobIndex = 0x11,             // Titan Blob DB only
+  // Titan has been using 0x10 and 0x11 (originally kTypeColumnFamilyBlobIndex
+  // and kTypeBlobIndex) for its own purposes. However, Titan and RocksDB native
+  // Blob DB diverged in their implementations, and RocksDB sometimes try to
+  // parse kTypeBlobIndex regardless of whether it is a native Blob DB or
+  // stackable implementation, thus causing issues. To avoid potential
+  // conflicts, we are reassigning these two values to new types. So that Titan
+  // does not need to migrate existing use cases.
+  kTypeTitanColumnFamilyBlobIndex = 0x10,  // Titan Blob DB only
+  kTypeTitanBlobIndex = 0x11,              // Titan Blob DB only
   // When the prepared record is also persisted in db, we use a different
   // record. This is to ensure that the WAL that is generated by a WritePolicy
   // is not mistakenly read by another, which would result into data
@@ -68,7 +75,11 @@ enum ValueType : unsigned char {
   kTypeCommitXIDAndTimestamp = 0x15,  // WAL only
   kTypeWideColumnEntity = 0x16,
   kTypeColumnFamilyWideColumnEntity = 0x17,  // WAL only
-  kTypeBlobIndex = 0x18,                     // RocksDB native Blob DB only
+  // Moving kTypeColumnFamilyBlobIndex and kTypeBlobIndex (originally 0x10 and
+  // 0x11) to 0x18 and 0x19 respectively to avoid potential conflicts with
+  // Titan. See comments above for more details.
+  kTypeColumnFamilyBlobIndex = 0x18,  // RocksDB native Blob DB only
+  kTypeBlobIndex = 0x19,              // RocksDB native Blob DB only
   kTypeMaxValid,    // Should be after the last valid type, only used for
                     // validation
   kMaxValue = 0x7F  // Not used for storing records.
@@ -82,7 +93,8 @@ extern const ValueType kValueTypeForSeekForPrev;
 // (i.e. a type used in memtable skiplist and sst file datablock).
 inline bool IsValueType(ValueType t) {
   return t <= kTypeMerge || kTypeSingleDeletion == t || kTypeBlobIndex == t ||
-         kTypeDeletionWithTimestamp == t || kTypeWideColumnEntity == t;
+         kTypeDeletionWithTimestamp == t || kTypeWideColumnEntity == t ||
+         kTypeTitanBlobIndex == t;
 }
 
 // Checks whether a type is from user operation
